v0.6.0: json parameter to get JSON for ad-hoc cmds
This adds a json parameter to ansible.command and ansible.command_local.
Using this parameter ensures that the stdout from a command is valid
JSON. This is especially helpful to make variables/facts from the setup
module (for example) available in an st2 workflow.

To make this work, we depend on ansible's --tree parameter to save the
output for each node in the 'tree' directory. We ignore all of ansible's
stdout, and instead feed the contents of all of the tree/node files in
one big json object (key is node name, and value is the output).

Bump to 0.6.0

Signed-off-by: Jacob Floyd <cognifloyd@gmail.com>