Add tests for extra_vars parsing
Also, based on tests, make it actually work.

The extra_vars.yaml fixture-based tests were written based on the master
code base. One test had to change based on my order-preserving change,
and the expected results are commented out to show the difference
between the old output and the new. The json and complex tests ensure
that the new json functionality works as expected.

Signed-off-by: Jacob Floyd <cognifloyd@gmail.com>